Place Elixir into local path when installing
This removes the need to add the `$HOME/elixir/bin` directory to the
PATH. It will just work thanks to symlinks.

This also fixes an odd issue when building libnl where Buildroot's host
path would get dropped when running commands. This broke the calls to
run `flex` and `bison`, but it could have been others.